x-pack/filebeat/input/netflow/decoder/...: move testing helpers out of exported package
Including the testing helpers in the template package results in
unnecessarily polluting the filebeat binary with the testing package.
So move it into its own package that is only imported by _test.go files.
