+++ x-pack/filebeat/input/netflow/decoder/ipfix/decoder_test.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/elastic/beats/v7/x-pack/filebeat/input/netflow/decoder/fields"
 	"github.com/elastic/beats/v7/x-pack/filebeat/input/netflow/decoder/template"
+	"github.com/elastic/beats/v7/x-pack/filebeat/input/netflow/decoder/template/tmpltest"
 	"github.com/elastic/beats/v7/x-pack/filebeat/input/netflow/decoder/test"
 	v9 "github.com/elastic/beats/v7/x-pack/filebeat/input/netflow/decoder/v9"
 )
@@ -146,7 +147,7 @@ func TestDecoderV9_ReadFields(t *testing.T) {
 			assert.Equal(t, tc.expected.Length, record.Length)
 			assert.Equal(t, tc.expected.VariableLength, record.VariableLength)
 			assert.Equal(t, tc.expected.ID, record.ID)
-			template.AssertFieldsEquals(t, tc.expected.Fields, record.Fields)
+			tmpltest.AssertFieldsEquals(t, tc.expected.Fields, record.Fields)
 		})
 	}
 }
@@ -267,7 +268,7 @@ func TestReadOptionsTemplateFlowSet(t *testing.T) {
 			assert.Equal(t, tc.err, err)
 			if assert.Len(t, templates, len(tc.expected)) {
 				for idx := range tc.expected {
-					template.AssertTemplateEquals(t, tc.expected[idx], templates[idx])
+					tmpltest.AssertTemplateEquals(t, tc.expected[idx], templates[idx])
 				}
 			}
 		})
@@ -340,7 +341,7 @@ func TestReadRecordTemplateFlowSet(t *testing.T) {
 			assert.Equal(t, tc.err, err)
 			if assert.Len(t, templates, len(tc.expected)) {
 				for idx := range tc.expected {
-					template.AssertTemplateEquals(t, tc.expected[idx], templates[idx])
+					tmpltest.AssertTemplateEquals(t, tc.expected[idx], templates[idx])
 				}
 			}
 		})
